First day of working on FragCraft
This is the first day, and i started up by finding some cool assets i like, basically imported a map that reminds me alot of an old CS map, i dont remember the name of, but its really cool. Then i imported a AK and some Particle effects for the gun aswell.

I implemented following mechanics today:

-FPS Movement
-FPS Camera
-Did some RayCast training and implemented a shooting mechanic
     -Using the RayCast, i inserted Particles at the end of the Ray,
       which got pretty cool :D

Plans for the following day will be making a simple prototype AI, and develop a NavMesh for that. Then when that is up and working, i will add weapons and event activated attacks to the AIs. Far in the future, maybe i will implement some Warcraft classes to not only the player but to the bots aswell, but that not on my Core Concept, so that is not a priority right now.