Added Sound VFX, Gun Materials and Animation Rigging
Sound VFX for Gun shooting and reloading, currently only for a AK47.

For the AK47 i added a Gold Material for everything except the magazine, which have a dotted metal material, which contrast really well with the gold.

Also rigged my player controllers hands, to hold to places on the AK47, the right hand on the trigger and left hand to hold the weight of the barrel.

Next to add:

Weapon Sway and recoil.